requirements.txt
coreapi
django-filter
djangorestframework
mysqlclient